refactor(theme): into TS
